cleaned up parser and made it independent of mnist model
